backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0
python-decouple==3.8